florestad: Add lib.rs
This commit introduces a lib.rs for florestad. With this change, other softwares
can spawn a florestad process and run on the background.
