+++ crates/floresta/Cargo.toml
@@ -46,6 +46,7 @@ floresta = { version = "0.1.0", path = "../floresta", features = [
     "electrum-server",
     "watch-only-wallet",
 ] }
+futures = "0.3.30"
 
 [features]
 default = ["bitcoinconsensus", "electrum-server", "watch-only-wallet"]
